requirements.txt
pandas>=2.0
pyyaml
pyarrow
pydantic
numpy
scikit-learn
joblib
pytest
